requirements.txt
fastapi~=0.115.4
sqlalchemy~=2.0.31
uvicorn~=0.30.1
pydantic~=2.8.2
alembic~=1.13.2
python-dotenv~=1.0.1
requests~=2.32.3
asyncpg~=0.30.0